ch13/flask/dev/work/pyproject.toml
[tool.poetry]
name = "simpleweb"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0"
mysql-connector-python = "*"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
